requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
langchain==0.0.107
PyYAML==6.0
nltk==3.7
openai==0.27.0
layoutparser==0.3.4
transformers==4.26.1
unstructured==0.5.3
python-magic==0.4.27
pinecone-client==2.2.1
-e git+https://github.com/facebookresearch/detectron2.git@v0.6#egg=detectron2

